backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
slowapi==0.1.9
python-vininfo==1.8.0
beautifulsoup4==4.12.3
lxml==5.2.2
cachetools==5.3.3
tenacity==8.3.0
